feat: [#56025] use structured filters
Before this change the integration would convert it-system user-keys to
it-system UUIDs to filter it-users by the it-systems of interest.

However, newer versions of OS2mo support filtering directly by user-key
by letting OS2mo do the it-system user-key to UUID lookup.

This commit makes use of this functionality within OS2mo and should
servce to be more performant and at the same time simply the integration.